:recycle: simplify
